added strategies for messaging
added strategies for storing
