+++ README.md
@@ -14,7 +14,7 @@ An open-source chat application to learn real-time messaging.
 
 ```
 NEXT_PUBLIC_MESSAGING=NEXT_PUBLIC_MESSAGING[local|cloud]
-NEXT_PUBLIC_STORAGE=NEXT_PUBLIC_STORAGE[local|cloud]
+NEXT_PUBLIC_STORING=NEXT_PUBLIC_STORING[local|cloud]
 NEXT_PUBLIC_SUPABASE_URL=YOUR_NEXT_PUBLIC_SUPABASE_URL
 NEXT_PUBLIC_SUPABASE_ANON_KEY=YOUR_NEXT_PUBLIC_SUPABASE_ANON_KEY
 NEXT_PUBLIC_DATABASE_URL=YOUR_NEXT_PUBLIC_DATABASE_URL
@@ -43,7 +43,6 @@ cd scripts
 chmod +x db.sh
 ./db.sh
 ```
-
 
 ### Start Application
